Remove ansible vault
Prior to this change We use ansible vault to store konstructor dns key.
Everytime we rotate a key, ansible vault has to be updated. There is no
point of using the ansible vault only for one secret.

This change removes ansible vault from the provisioner and sets the
konstructor_api_key as environment variable.